More fs2 prefixes removed
