+++ modules/streams/src/main/scala/dev/profunktor/fs2redis/interpreter/pubsub/Publisher.scala
@@ -17,40 +17,22 @@
 package dev.profunktor.redis4cats.interpreter.pubsub
 
 import cats.effect.{ ConcurrentEffect, ContextShift, Sync }
-import cats.effect.concurrent.Ref
-import cats.syntax.all._
-import dev.profunktor.redis4cats.algebra.{ PubSubCommands, PubSubStats, SubscribeCommands }
+import cats.syntax.functor._
+import dev.profunktor.redis4cats.algebra.{ PubSubStats, PublishCommands }
 import dev.profunktor.redis4cats.domain.RedisChannel
-import dev.profunktor.redis4cats.interpreter.pubsub.internals.{ Fs2PubSubInternals, PubSubState }
 import dev.profunktor.redis4cats.streams.Subscription
-import dev.profunktor.redis4cats.effect.{ JRFuture, Log }
+import dev.profunktor.redis4cats.effect.JRFuture
 import fs2.Stream
 import io.lettuce.core.pubsub.StatefulRedisPubSubConnection
 
-class Fs2PubSubCommands[F[_]: ConcurrentEffect: ContextShift: Log, K, V](
-    state: Ref[F, PubSubState[F, K, V]],
-    subConnection: StatefulRedisPubSubConnection[K, V],
-    pubConnection: StatefulRedisPubSubConnection[K, V]
-) extends PubSubCommands[Stream[F, ?], K, V] {
+class Publisher[F[_]: ConcurrentEffect: ContextShift, K, V](pubConnection: StatefulRedisPubSubConnection[K, V])
+    extends PublishCommands[Stream[F, ?], K, V] {
 
-  private[redis4cats] val subCommands: SubscribeCommands[Stream[F, ?], K, V] =
-    new Fs2Subscriber[F, K, V](state, subConnection)
-  private[redis4cats] val pubSubStats: PubSubStats[Stream[F, ?], K] = new Fs2PubSubStats(pubConnection)
-
-  override def subscribe(channel: RedisChannel[K]): Stream[F, V] =
-    subCommands.subscribe(channel)
-
-  override def unsubscribe(channel: RedisChannel[K]): Stream[F, Unit] =
-    subCommands.unsubscribe(channel)
+  private[redis4cats] val pubSubStats: PubSubStats[Stream[F, ?], K] = new LivePubSubStats(pubConnection)
 
   override def publish(channel: RedisChannel[K]): Stream[F, V] => Stream[F, Unit] =
     _.evalMap { message =>
-      val getOrCreateTopicListener = Fs2PubSubInternals[F, K, V](state, subConnection)
-      for {
-        st <- state.get
-        _ <- getOrCreateTopicListener(channel)(st)
-        _ <- JRFuture { Sync[F].delay(pubConnection.async().publish(channel.value, message)) }
-      } yield ()
+      JRFuture { Sync[F].delay(pubConnection.async().publish(channel.value, message)) }.void
     }
 
   override def pubSubChannels: Stream[F, List[K]] =
